TrailMate/requirements.txt
openai>=1.0.0
gradio>=4.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
requests>=2.31.0
